Move unit and integration tests to separate directories
