+++ tests/integration/seedDatabase.ts
@@ -1,10 +1,10 @@
 import { mkdirSync, removeSync, existsSync, writeFileSync } from 'fs-extra';
 import { MongoClient } from 'mongodb';
 
-import { DeepPartial, AppConfig, defaultConfig } from '../src/config';
-import { seedDatabase } from '../src/index';
-import { DatabaseConnector } from '../src/DatabaseConnector';
-import { Database } from '../src/Database';
+import { DeepPartial, AppConfig, defaultConfig } from '../../src/config';
+import { seedDatabase } from '../../src/index';
+import { DatabaseConnector } from '../../src/DatabaseConnector';
+import { Database } from '../../src/Database';
 
 const TEMP_DIRECTORY_PATH = __dirname + '/tempIntegration';
